Rename test and extract fixture (in `tests/v2/04_runner`)
